+++ frontend/src/components/Layout/Nav/Nav.tsx
@@ -1,7 +1,12 @@
+import AddBoxIcon from '@mui/icons-material/AddBox'
 import { Button, Grid, Typography } from '@mui/material'
+import { styled } from '@mui/material/styles'
 import { BarButton } from '../../BarButton/BarButton'
-import { StyledIcon, buttonSx } from './constants'
 import { NavBarButton } from './NavBarButton/NavBarButton'
+
+const StyledIcon = styled(AddBoxIcon)(({ theme }) => ({
+  marginRight: theme.spacing(2)
+}))
 
 export const Nav = () => {
   return (
@@ -12,7 +17,27 @@ export const Nav = () => {
       <NavBarButton to="/tactical" label="TACTICAL" />
       <Grid>
         <BarButton>
-          <Button sx={buttonSx} variant="contained" size="small" disabled>
+          <Button
+            sx={{
+              px: 4,
+              backgroundColor: '#fff',
+              color: '#000',
+              position: 'relative',
+              zIndex: 2,
+              boxShadow: 'none',
+              border: 'none',
+              clipPath: 'none',
+              lineHeight: 1,
+              '&:hover': {
+                backgroundColor: 'transparent',
+                color: '#fff',
+                boxShadow: 'none'
+              }
+            }}
+            variant="contained"
+            size="small"
+            disabled
+          >
             <StyledIcon fontSize="small" />
             RESEARCH
           </Button>
@@ -20,7 +45,27 @@ export const Nav = () => {
       </Grid>
       <Grid>
         <BarButton>
-          <Button sx={buttonSx} variant="contained" size="small" disabled>
+          <Button
+            sx={{
+              px: 4,
+              backgroundColor: '#fff',
+              color: '#000',
+              position: 'relative',
+              zIndex: 2,
+              boxShadow: 'none',
+              border: 'none',
+              clipPath: 'none',
+              lineHeight: 1,
+              '&:hover': {
+                backgroundColor: 'transparent',
+                color: '#fff',
+                boxShadow: 'none'
+              }
+            }}
+            variant="contained"
+            size="small"
+            disabled
+          >
             <StyledIcon fontSize="small" />
             GALAXY
           </Button>